rename ui module for app
