+++ src/app/component/badge.rs
