+++ src/app/component/icon.rs
