+++ src/app/component/text.rs
