+++ src/app/view/spend_transaction.rs
@@ -3,14 +3,14 @@ use bitcoin::util::psbt::PartiallySignedTransaction as Psbt;
 use iced::{scrollable, Align, Column, Container, Element, Length, Row, TextInput};
 
 use crate::{
-    revaultd::model,
-    ui::{
+    app::{
         component::{badge, button, card, scroll, text, ContainerBackgroundStyle},
         error::Error,
         menu::Menu,
         message::{Message, SpendTxMessage},
         view::{manager::spend_tx_with_feerate_view, Context},
     },
+    revaultd::model,
 };
 
 #[derive(Debug)]
