--- src/ui/mod.rs
@@ -1,8 +1,0 @@
-pub mod app;
-mod component;
-pub mod config;
-mod error;
-mod menu;
-mod message;
-mod state;
-mod view;
